pyproject.toml
[tool.poetry]
name = "langgraph_func"
version = "0.3.1"
description = "LangGraph Azure Function Agents"
authors = ["job van Creij jobvancreij@gmail.com"]
packages = [
    { include = "langgraph_func" }
]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
azure-functions = "*"
pydantic = "*"
langgraph = "*"
openapi-schema-pydantic = "*"
requests = "*"
dacite = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.docs.dependencies]
sphinx = "*"
myst-parser = "*"
toml = "*"

[tool.poetry.urls]
Documentation = "https://jobaibv.github.io/langgraph_func/"